CAMEL-19632: Fix split package in camel-xml-jaxp and camel-xml-japx-util
