+++ core/camel-xml-jaxp-util/src/main/java/org/apache/camel/util/xml/pretty/XmlPrettyPrinter.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.util.xml;
+package org.apache.camel.util.xml.pretty;
 
 import java.io.ByteArrayInputStream;
