--- core/camel-xml-jaxp-util/src/test/resources/org/apache/camel/util/camel-context.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Licensed to the Apache Software Foundation (ASF) under one or more
-    contributor license agreements.  See the NOTICE file distributed with
-    this work for additional information regarding copyright ownership.
-    The ASF licenses this file to You under the Apache License, Version 2.0
-    (the "License"); you may not use this file except in compliance with
-    the License.  You may obtain a copy of the License at
-
-         http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xsi:schemaLocation="
-       http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
-       http://camel.apache.org/schema/spring http://camel.apache.org/schema/spring/camel-spring.xsd">
-
-  <!-- notice Camel will only update the routes that has been changed, so you can edit either either route or both
-       and save the file, and Camel will update only what is required -->
-
-  <camelContext xmlns="http://camel.apache.org/schema/spring">
-
-    <route id="timer">
-      <from uri="timer:foo"/>
-      <!-- call the 2nd route -->
-      <to uri="direct:foo"/>
-      <!-- try to change me and save this file -->
-      <log message="You said: ${body}"/>
-    </route>
-
-    <route id="foo">
-      <from uri="direct:foo"/>
-      <!-- try to change me and save this file -->
-      <transform>
-        <constant>Hello World</constant>
-      </transform>
-    </route>
-
-  </camelContext>
-
-</beans>
